% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_differential_expression}
\alias{run_differential_expression}
\title{Helper function for rendering the differential expression RMarkdown document}
\usage{
run_differential_expression(
  path_config_json,
  path_salmon_counts = "",
  out_path,
  count_data = NULL,
  save_csv = TRUE,
  save_excel = TRUE,
  save_deseq_rds = TRUE,
  biomart_version = 97,
  biomart_attributes = "none",
  clean_output = TRUE
)
}
\arguments{
\item{path_config_json}{Path to experiment design json file}

\item{path_salmon_counts}{Path to salmon counts file}

\item{out_path}{Path for output}

\item{count_data}{Count data dataframe. Replaces path_salmon_counts}

\item{save_csv}{Output differentially expressed entries as csv file}

\item{save_excel}{Output differentially expressed entries as excel file}

\item{save_deseq_rds}{Output DESeq2 result object as rds file}

\item{biomart_version}{Specify Ensembl version. Only required if biomart_attributes are defined}

\item{biomart_attributes}{Attach biomart attributes to the resulting table}

\item{clean_output}{Clean intermediate rmarkdown render files}
}
\description{
Helper function for rendering the differential expression RMarkdown document
}
