% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{goterm_analysis_of_all_comparisons}
\alias{goterm_analysis_of_all_comparisons}
\title{Create GO-term analysis for all comparisons of deseq2_diff.csv (created by nfRNAseqDESeq2)}
\usage{
goterm_analysis_of_all_comparisons(
  deseq2_diff_path,
  out_path,
  simplify_ontologies = TRUE,
  do_gse = TRUE
)
}
\arguments{
\item{deseq2_diff_path}{Path to deseq2_diff.csv created by nfRNAseqDESeq2}

\item{out_path}{Path to output files}

\item{simplify_ontologies}{See mygo::createHTMLReport}

\item{do_gse}{See mygo::createHTMLReport}
}
\description{
Create GO-term analysis for all comparisons of deseq2_diff.csv (created by nfRNAseqDESeq2)
}
\examples{
  goterm_analysis_of_all_comparisons(
    deseq2_diff_path = "/beegfs/scratch/bruening_scratch/pklemm/2019-11-sinika-rnaseq/analysis/results/DESeq2/deseq2_diff.csv",
    out_path = "/beegfs/scratch/bruening_scratch/pklemm/2019-11-sinika-rnaseq/analysis/results/goterm-analysis"
  )
}
