% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_differential_expression}
\alias{run_differential_expression}
\title{Helper function for rendering the differential expression RMarkdown document}
\usage{
run_differential_expression(
  path_config_json,
  out_path,
  count_data,
  count_data_normalized,
  save_intermediate = TRUE,
  ensembl_version = 101,
  biomart_attributes = "none",
  clean_output = TRUE,
  minimum_count = 0,
  minimum_padj = 0.05,
  biotypes_filter = "protein_coding"
)
}
\arguments{
\item{path_config_json}{Path to experiment design json file}

\item{out_path}{Path for output}

\item{count_data}{Count data. Replaces path_salmon_counts}

\item{count_data_normalized}{Normalized count dataframe. Replaces path_salmon_tpm}

\item{save_intermediate}{Output intermediate files that can be used for debugging/custom analyses}

\item{ensembl_version}{Specify Ensembl version. Only required if biomart_attributes are defined}

\item{biomart_attributes}{Attach biomart attributes to the resulting table}

\item{clean_output}{Clean intermediate rmarkdown render files}

\item{minimum_count}{For each gene at least one sample must equal or larger than this value to be included in the analysis}

\item{minimum_padj}{Minimum padj value for a gene to be considered differentially expressed}

\item{biotypes_filter}{Filter genes for these biotypes}
}
\description{
Helper function for rendering the differential expression RMarkdown document
}
